+++ libs/TYTileMapSDK.framework/Versions/A/Headers/TYTileMapSDK.h
@@ -9,10 +9,5 @@
 #import <TYTileMapSDK/TYTiledLayer.h>
 #import <TYTileMapSDK/TYTiledManager.h>
 
-
-/**
- * 1.2    优化瓦片加载链接
- * 1.1    优化瓦片加载
- *        修复tileForFloor传入纯数字
- */
-#define TYTileMapSDKVersionNumber 1.2
+//! Project version number for TYTileMapSDK.
+#define TYTileMapSDKVersionNumber 1.1